refactor: split meshV2 scanning step into separate component
Extract meshV2-specific scanning step UI into its own file to keep
the upstream scanning-step.jsx clean and make future meshV2 UI
changes easier.

- Create mesh-v2-scanning-step.jsx with HiraganaNameSearch and
  MeshV2ScanningStep components (Smalruby-specific file)
- Restore scanning-step.jsx to upstream-compatible version (remove
  all meshV2 markers, extensionId checks, and name search props)
- Update scanning-step container to render MeshV2ScanningStep when
  extensionId is 'meshV2', standard ScanningStep otherwise
- Update tests to import from mesh-v2-scanning-step.jsx

Refs #342
